more things, include C++ traits, multiprocess and thread, threadpool, TCP-IP network, and so on...

+++ concurrency/Threading_In_CPlusPlus/1.thread/intro.cpp
@@ -36,6 +36,6 @@ int main() {
 
     cout << "OddSum : " << OddSum << endl;
     cout << "EvenSum: " << EvenSum << endl;
-    cout << "Sec: " << duration.count() / 1000000 << endl;
+    cout << "Sec: " << (double)duration.count() / 1000000 << endl;
     return 0;
 }
